requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pinecone-client==4.1.1
openai==1.35.3
python-dotenv==1.0.1
pydantic==2.7.4
